// frontend/src/api/model/prescriptionKind.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type PrescriptionKind = typeof PrescriptionKind[keyof typeof PrescriptionKind];


export const PrescriptionKind = {
  MEDICATION: 'MEDICATION',
  PRACTICE: 'PRACTICE',
} as const;